Switch Mocha tests from TS to JS, add RemoteStorage mocha suite
