+++ package.json
@@ -13,8 +13,8 @@
   "homepage": "https://remotestorage.io",
   "scripts": {
     "test": "tsc && bash scripts/test-all.sh",
-    "test:mocha": "mocha test/unit/*.test.ts",
-    "test:watch": "mocha test/unit/*.test.ts --watch --reporter dot",
+    "test:mocha": "mocha test/unit/*.test.mjs",
+    "test:watch": "mocha test/unit/*.test.mjs --watch --reporter dot",
     "lint": "eslint --ext=ts src/",
     "lint:quiet": "eslint --quiet --ext=ts src/",
     "lint:specs": "eslint test/unit/*.ts",
