+++ test/unit/remotestorage-suite.js
@@ -171,81 +171,6 @@ define(['require', 'tv4', './build/eventhandling', './build/util'],
       },
 
       {
-        desc: "#setApiKeys initializes the configured backend when it's not initialized yet",
-        run: function(env, test) {
-          Dropbox._rs_init = test.done;
-
-          env.rs.setApiKeys({ dropbox: 'testkey' });
-        }
-      },
-
-      {
-        desc: "#setApiKeys reinitializes the configured backend when the key changed",
-        run: function(env, test) {
-          env.rs.dropbox = {
-            clientId: 'old key'
-          };
-
-          Dropbox._rs_init = test.done;
-
-          env.rs.setApiKeys({ dropbox: 'new key' });
-        }
-      },
-
-      {
-        desc: "#setApiKeys does not reinitialize the configured backend when key didn't change",
-        run: function(env, test) {
-          env.rs.dropbox = {
-            clientId: 'old key'
-          };
-
-          Dropbox._rs_init = function(remoteStorage) {
-              test.fail('Backend got reinitialized again although the key did not change.');
-            };
-
-          env.rs.setApiKeys({ dropbox: 'old key' });
-          test.done();
-        }
-      },
-
-      {
-        desc: "#setApiKeys allows setting values for 'googledrive' and 'dropbox'",
-        run: function(env, test) {
-          env.rs.setApiKeys({
-            dropbox: '123abc',
-            googledrive: '456def'
-          });
-
-          test.assert(env.rs.apiKeys['dropbox'].appKey, '123abc');
-          test.assert(env.rs.apiKeys['googledrive'].clientId, '456def');
-          test.assert(env.rs.dropbox.clientId, '123abc');
-          test.assert(env.rs.googledrive.clientId, '456def');
-        }
-      },
-
-      {
-        desc: "#setApiKeys returns false when receiving invalid config",
-        run: function(env, test) {
-          test.assert(env.rs.setApiKeys({ icloud: '123abc' }), false);
-        }
-      },
-
-      {
-        desc: "#setApiKeys clears config when receiving null values",
-        run: function(env, test) {
-          env.rs.setApiKeys({
-            dropbox: null,
-            googledrive: null
-          });
-
-          test.assertType(env.rs.apiKeys['dropbox'], 'undefined');
-          test.assertType(env.rs.apiKeys['googledrive'], 'undefined');
-          test.assert(env.rs.dropbox.clientId, 'null');
-          test.assert(env.rs.googledrive.clientId, 'null');
-        }
-      },
-
-      {
         desc: "#connect throws unauthorized when userAddress doesn't contain an @ or URL",
         run: function(env, test) {
           env.rs.on('error', function(e) {
